feat: Add MyHttpClient and MicroServiceHttpClient classes
This commit adds two new classes, `MyHttpClient` and `MicroServiceHttpClient`, to the codebase.

`MyHttpClient` is responsible for fetching data from an API using the Symfony HttpClient component. It takes an access token as a constructor argument and uses it to authenticate the request. The fetched data is returned as an array.

`MicroServiceHttpClient` is similar to `MyHttpClient`, but it is specifically designed to fetch data from a microservice API. It also takes an access token as a constructor argument and uses it for authentication. The fetched data is returned as an array.

These classes will be used to improve the data fetching capabilities of the application.